upgrade anakin, fix log parser stream detach, update chalk module

+++ test/lib/spawn.js
@@ -1,16 +1,36 @@
-const nativeSpawn = require('child_process').spawn;
+const os = require('os');
+const childProcess = require('child_process');
+const nativeSpawn = childProcess.spawn;
+const exec = childProcess.exec;
 
-const children = [];
+let children = [];
 const spawn = (...rest) => {
   const child = nativeSpawn(...rest);
   children.push(child);
   return child;
 };
 
-const anakin = () => {
-  children.forEach((child) => {
-    child.kill('SIGINT');
-  });
+const anakin = (done) => {
+  return Promise.all(children
+    .map((child) => new Promise((resolve) => {
+      if (child.killed || [0, 1].includes(child.exitCode)) {
+        resolve();
+        return;
+      }
+      if (os.platform() === 'win32') {
+        child.on('exit', () => setTimeout(() => resolve(), 500))
+        exec('taskkill /pid ' + child.pid + ' /T /F');
+      } else {
+        child.kill('SIGINT');
+        resolve();
+      }
+    })))
+    .then(() => {
+      children = [];
+      if(typeof done === 'function') {
+        done();
+      }
+    });
 };
 
 process.on('error', () => {
